Add solution for problem `C - Superior Periodic Subarrays` Link: https://codeforces.com/contest/582/problem/C Tags: number theory, *2400 Ref: https://codeforces.com/contest/582/submission/13383105

+++ codeforces/582/C.cpp
@@ -18,7 +18,7 @@ const int MN = 200111;
 
 int a[MN], n;
 int max_rem[MN], good[MN*2];
-int x[MN];
+ll x[MN];
 int cache_gcd[MN];
 
 ll get(int d) {
